Update xlsx.js to allow cmpat with non office generated files.
Some xlsx files generated by 3rd party applications do not come with valid property data in the document formatting. The changes to xlsx.js in this patch seek to allow compatibility with these out of spec files while informing the user of the invalid data.

A simple 'fix' to a problem that shouldn't exist but for some reason 3rd party applications prefer to discard design specifications.